+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 Version 3.4.0 (2019-09-XX)
 ==========================
 - Add support for pyarrow 0.14.1
+- Use urlquote for faster quoting/unquoting
+
 
 Version 3.3.0 (2019-08-15)
 ==========================